Type Definition packed_simd::f32x16

type f32x16 = Simd<[f32; 16]>;

A 512-bit vector with 16 f32 lanes.

Methods

impl f32x16

pub const fn new(
x0: f32,
x1: f32,
x2: f32,
x3: f32,
x4: f32,
x5: f32,
x6: f32,
x7: f32,
x8: f32,
x9: f32,
x10: f32,
x11: f32,
x12: f32,
x13: f32,
x14: f32,
x15: f32
) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: f32) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> f32

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> f32

Extracts the value at index.

Precondition

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub fn replace(self, index: usize, new_value: f32) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub unsafe fn replace_unchecked(self, index: usize, new_value: f32) -> Self

Returns a new vector where the value at index is replaced by new_value.

Precondition

If index >= Self::lanes() the behavior is undefined.

impl f32x16

pub fn min(self, x: Self) -> Self

Minimum of two vectors.

Returns a new vector containing the minimum value of each of the input vector lanes.

pub fn max(self, x: Self) -> Self

Maximum of two vectors.

Returns a new vector containing the maximum value of each of the input vector lanes.

impl f32x16

pub fn sum(self) -> f32

Horizontal sum of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 + x1) + (x2 + x3)) + ((x4 + x5) + (x6 + x7))

If one of the vector element is NaN the reduction returns NaN. The resulting NaN is not required to be equal to any of the NaNs in the vector.

pub fn product(self) -> f32

Horizontal product of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 * x1) * (x2 * x3)) * ((x4 * x5) * (x6 * x7))

If one of the vector element is NaN the reduction returns NaN. The resulting NaN is not required to be equal to any of the NaNs in the vector.

impl f32x16

pub fn max_element(self) -> f32

Largest vector element value.

pub fn min_element(self) -> f32

Smallest vector element value.

impl f32x16

pub fn from_slice_aligned(slice: &[f32]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn from_slice_unaligned(slice: &[f32]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn from_slice_aligned_unchecked(slice: &[f32]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn from_slice_unaligned_unchecked(slice: &[f32]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl f32x16

pub fn write_to_slice_aligned(self, slice: &mut [f32])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn write_to_slice_unaligned(self, slice: &mut [f32])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn write_to_slice_aligned_unchecked(self, slice: &mut [f32])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn write_to_slice_unaligned_unchecked(self, slice: &mut [f32])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl f32x16

pub fn abs(self) -> Self

Absolute value.

impl f32x16

pub fn cos(self) -> Self

Cosine.

pub fn cos_pi(self) -> Self

Cosine of self * PI.

impl f32x16

pub fn exp(self) -> Self

Returns the exponential function of self: e^(self).

impl f32x16

pub fn ln(self) -> Self

Returns the natural logarithm of self.

impl f32x16

pub fn mul_add(self, y: Self, z: Self) -> Self

Fused multiply add: self * y + z

impl f32x16

pub fn mul_adde(self, y: Self, z: Self) -> Self

Fused multiply add estimate: ~= self * y + z

While fused multiply-add (fma) has infinite precision, mul_adde has at worst the same precision of a multiply followed by an add. This might be more efficient on architectures that do not have an fma instruction.

impl f32x16

pub fn powf(self, x: Self) -> Self

Raises self number to the floating point power of x.

impl f32x16

pub fn recpre(self) -> Self

Reciprocal estimate: ~= 1. / self.

FIXME: The precision of the estimate is currently unspecified.

impl f32x16

pub fn rsqrte(self) -> Self

Reciprocal square-root estimate: ~= 1. / self.sqrt().

FIXME: The precision of the estimate is currently unspecified.

impl f32x16

pub fn sin(self) -> Self

Sine.

pub fn sin_pi(self) -> Self

Sine of self * PI.

pub fn sin_cos_pi(self) -> (Self, Self)

Sine and cosine of self * PI.

impl f32x16

pub fn sqrt(self) -> Self

impl f32x16

pub fn sqrte(self) -> Self

Square-root estimate.

FIXME: The precision of the estimate is currently unspecified.

impl f32x16

pub fn eq(self, other: Self) -> m32x16

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m32x16

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m32x16

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m32x16

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m32x16

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m32x16

Lane-wise greater-than-or-equals comparison.