Type Definition packed_simd::i8x16

type i8x16 = Simd<[i8; 16]>;

A 128-bit vector with 16 i8 lanes.

Methods

impl i8x16

pub const fn new(
x0: i8,
x1: i8,
x2: i8,
x3: i8,
x4: i8,
x5: i8,
x6: i8,
x7: i8,
x8: i8,
x9: i8,
x10: i8,
x11: i8,
x12: i8,
x13: i8,
x14: i8,
x15: i8
) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: i8) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> i8

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> i8

Extracts the value at index.

Precondition

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub fn replace(self, index: usize, new_value: i8) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub unsafe fn replace_unchecked(self, index: usize, new_value: i8) -> Self

Returns a new vector where the value at index is replaced by new_value.

Precondition

If index >= Self::lanes() the behavior is undefined.

impl i8x16

pub fn rotate_left(self, n: i8x16) -> i8x16

Shifts the bits of each lane to the left by the specified amount in the corresponding lane of n, wrapping the truncated bits to the end of the resulting integer.

Note: this is neither the same operation as << nor equivalent to slice::rotate_left.

pub fn rotate_right(self, n: i8x16) -> i8x16

Shifts the bits of each lane to the right by the specified amount in the corresponding lane of n, wrapping the truncated bits to the beginning of the resulting integer.

Note: this is neither the same operation as << nor equivalent to slice::rotate_left.

impl i8x16

pub fn min(self, x: Self) -> Self

Minimum of two vectors.

Returns a new vector containing the minimum value of each of the input vector lanes.

pub fn max(self, x: Self) -> Self

Maximum of two vectors.

Returns a new vector containing the maximum value of each of the input vector lanes.

impl i8x16

pub fn wrapping_sum(self) -> i8

Horizontal wrapping sum of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 + x1) + (x2 + x3)) + ((x4 + x5) + (x6 + x7))

If an operation overflows it returns the mathematical result modulo 2^n where n is the number of times it overflows.

pub fn wrapping_product(self) -> i8

Horizontal wrapping product of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 * x1) * (x2 * x3)) * ((x4 * x5) * (x6 * x7))

If an operation overflows it returns the mathematical result modulo 2^n where n is the number of times it overflows.

impl i8x16

pub fn max_element(self) -> i8

Largest vector element value.

pub fn min_element(self) -> i8

Smallest vector element value.

impl i8x16

pub fn and(self) -> i8

Lane-wise bitwise and of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn or(self) -> i8

Lane-wise bitwise or of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn xor(self) -> i8

Lane-wise bitwise xor of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

impl i8x16

pub fn from_slice_aligned(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn from_slice_unaligned(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn from_slice_aligned_unchecked(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn from_slice_unaligned_unchecked(slice: &[i8]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl i8x16

pub fn write_to_slice_aligned(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn write_to_slice_unaligned(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn write_to_slice_aligned_unchecked(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn write_to_slice_unaligned_unchecked(self, slice: &mut [i8])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl i8x16

pub fn swap_bytes(self) -> Self

Reverses the byte order of the vector.

pub fn to_le(self) -> Self

Converts self to little endian from the target's endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

pub fn to_be(self) -> Self

Converts self to big endian from the target's endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

impl i8x16

pub fn eq(self, other: Self) -> m8x16

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m8x16

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m8x16

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m8x16

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m8x16

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m8x16

Lane-wise greater-than-or-equals comparison.

impl i8x16

pub fn partial_lex_ord(&self) -> LexicographicallyOrdered<i8x16>

Returns a wrapper that implements PartialOrd.

impl i8x16

pub fn lex_ord(&self) -> LexicographicallyOrdered<i8x16>

Returns a wrapper that implements Ord.