Type Definition packed_simd::m32x4

type m32x4 = Simd<[m32; 4]>;

A 128-bit vector mask with 4 m32 lanes.

Methods

impl m32x4

pub const fn new(x0: bool, x1: bool, x2: bool, x3: bool) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: bool) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> bool

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> bool

Extracts the value at index.

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub fn replace(self, index: usize, new_value: bool) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"]
pub unsafe fn replace_unchecked(self, index: usize, new_value: bool) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

impl m32x4

pub fn and(self) -> bool

Lane-wise bitwise and of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn or(self) -> bool

Lane-wise bitwise or of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn xor(self) -> bool

Lane-wise bitwise xor of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

impl m32x4

pub fn all(self) -> bool

Are all vector lanes true?

pub fn any(self) -> bool

Is any vector lane true?

pub fn none(self) -> bool

Are all vector lanes false?

impl m32x4

pub fn eq(self, other: Self) -> m32x4

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m32x4

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m32x4

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m32x4

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m32x4

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m32x4

Lane-wise greater-than-or-equals comparison.

impl m32x4

pub fn select<T>(self, a: Simd<T>, b: Simd<T>) -> Simd<T> where
T: SimdArray<NT = <[m32; 4] as SimdArray>::NT>,

Selects elements of a and b using mask.

The lanes of the result for which the mask is true contain the values of a. The remaining lanes contain the values of b.

impl m32x4

pub fn partial_lex_ord(&self) -> LexicographicallyOrdered<m32x4>

Returns a wrapper that implements PartialOrd.

impl m32x4

pub fn lex_ord(&self) -> LexicographicallyOrdered<m32x4>

Returns a wrapper that implements Ord.