Type Definition packed_simd::u128x4

type u128x4 = Simd<[u128; 4]>;

A 512-bit vector with 4 u128 lanes.

Methods

impl u128x4

pub const fn new(x0: u128, x1: u128, x2: u128, x3: u128) -> Self

Creates a new instance with each vector elements initialized with the provided values.

pub const fn lanes() -> usize

Returns the number of vector lanes.

pub const fn splat(value: u128) -> Self

Constructs a new instance with each element initialized to value.

pub fn extract(self, index: usize) -> u128

Extracts the value at index.

Panics

If index >= Self::lanes().

pub unsafe fn extract_unchecked(self, index: usize) -> u128

Extracts the value at index.

Precondition

If index >= Self::lanes() the behavior is undefined.

#[must_use = "replace does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"] pub fn replace(self, index: usize, new_value: u128) -> Self

Returns a new vector where the value at index is replaced by new_value.

Panics

If index >= Self::lanes().

#[must_use = "replace_unchecked does not modify the original value - it returns a new vector with the value at `index` replaced by `new_value`d"] pub unsafe fn replace_unchecked(self, index: usize, new_value: u128) -> Self

Returns a new vector where the value at index is replaced by new_value.

Precondition

If index >= Self::lanes() the behavior is undefined.

impl u128x4

pub fn rotate_left(self, n: u128x4) -> u128x4

Shifts the bits of each lane to the left by the specified amount in the corresponding lane of n, wrapping the truncated bits to the end of the resulting integer.

Note: this is neither the same operation as << nor equivalent to slice::rotate_left.

pub fn rotate_right(self, n: u128x4) -> u128x4

Shifts the bits of each lane to the right by the specified amount in the corresponding lane of n, wrapping the truncated bits to the beginning of the resulting integer.

Note: this is neither the same operation as << nor equivalent to slice::rotate_left.

impl u128x4

pub fn min(self, x: Self) -> Self

Minimum of two vectors.

Returns a new vector containing the minimum value of each of the input vector lanes.

pub fn max(self, x: Self) -> Self

Maximum of two vectors.

Returns a new vector containing the maximum value of each of the input vector lanes.

impl u128x4

pub fn wrapping_sum(self) -> u128

Horizontal wrapping sum of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 + x1) + (x2 + x3)) + ((x4 + x5) + (x6 + x7))

If an operation overflows it returns the mathematical result modulo 2^n where n is the number of times it overflows.

pub fn wrapping_product(self) -> u128

Horizontal wrapping product of the vector elements.

The intrinsic performs a tree-reduction of the vector elements. That is, for an 8 element vector:

((x0 * x1) * (x2 * x3)) * ((x4 * x5) * (x6 * x7))

If an operation overflows it returns the mathematical result modulo 2^n where n is the number of times it overflows.

impl u128x4

pub fn max_element(self) -> u128

Largest vector element value.

pub fn min_element(self) -> u128

Smallest vector element value.

impl u128x4

pub fn and(self) -> u128

Lane-wise bitwise and of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn or(self) -> u128

Lane-wise bitwise or of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

pub fn xor(self) -> u128

Lane-wise bitwise xor of the vector elements.

Note: if the vector has one lane, the first element of the vector is returned.

impl u128x4

pub fn from_slice_aligned(slice: &[u128]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn from_slice_unaligned(slice: &[u128]) -> Self

Instantiates a new vector with the values of the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn from_slice_aligned_unchecked(slice: &[u128]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn from_slice_unaligned_unchecked(slice: &[u128]) -> Self

Instantiates a new vector with the values of the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl u128x4

pub fn write_to_slice_aligned(self, slice: &mut [u128])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary.

pub fn write_to_slice_unaligned(self, slice: &mut [u128])

Writes the values of the vector to the slice.

Panics

If slice.len() < Self::lanes().

pub unsafe fn write_to_slice_aligned_unchecked(self, slice: &mut [u128])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() or &slice[0] is not aligned to an align_of::<Self>() boundary, the behavior is undefined.

pub unsafe fn write_to_slice_unaligned_unchecked(self, slice: &mut [u128])

Writes the values of the vector to the slice.

Precondition

If slice.len() < Self::lanes() the behavior is undefined.

impl u128x4

pub fn swap_bytes(self) -> Self

Reverses the byte order of the vector.

pub fn to_le(self) -> Self

Converts self to little endian from the target's endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

pub fn to_be(self) -> Self

Converts self to big endian from the target's endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

pub fn from_le(x: Self) -> Self

Converts a vector from little endian to the target's endianness.

On little endian this is a no-op. On big endian the bytes are swapped.

pub fn from_be(x: Self) -> Self

Converts a vector from big endian to the target's endianness.

On big endian this is a no-op. On little endian the bytes are swapped.

impl u128x4

pub fn count_ones(self) -> Self

Returns the number of ones in the binary representation of the lanes of self.

pub fn count_zeros(self) -> Self

Returns the number of zeros in the binary representation of the lanes of self.

pub fn leading_zeros(self) -> Self

Returns the number of leading zeros in the binary representation of the lanes of self.

pub fn trailing_zeros(self) -> Self

Returns the number of trailing zeros in the binary representation of the lanes of self.

impl u128x4

pub fn shuffle1_dyn<I>(self, indices: I) -> Self where
Self: Shuffle1Dyn<Indices = I>,

Shuffle vector elements according to indices.

impl u128x4

pub fn eq(self, other: Self) -> m128x4

Lane-wise equality comparison.

pub fn ne(self, other: Self) -> m128x4

Lane-wise inequality comparison.

pub fn lt(self, other: Self) -> m128x4

Lane-wise less-than comparison.

pub fn le(self, other: Self) -> m128x4

Lane-wise less-than-or-equals comparison.

pub fn gt(self, other: Self) -> m128x4

Lane-wise greater-than comparison.

pub fn ge(self, other: Self) -> m128x4

Lane-wise greater-than-or-equals comparison.

impl u128x4

pub fn partial_lex_ord(&self) -> LexicographicallyOrdered<u128x4>

Returns a wrapper that implements PartialOrd.

impl u128x4

pub fn lex_ord(&self) -> LexicographicallyOrdered<u128x4>

Returns a wrapper that implements Ord.

impl u128x4

pub fn bitmask(self) -> u8

Creates a bitmask with the MSB of each vector lane.

If the vector has less than 8 lanes, the bits that do not correspond to any vector lanes are cleared.